requirements.txt
pymongo==4.2.0
beautifulsoup4==4.11.1
discord-py-interactions==4.3.1
requests==2.28.1
dnspython==2.2.1